Renames the GroupMembershipRoleDetails file
